move matrices up a level
