+++ .github/workflows/bottle.yml
@@ -7,24 +7,31 @@ on:
         type: boolean
         required: false
         default: false
+      platform:
+        required: true
+        type: string
 
 jobs:
+  get-platform:
+    runs-on: ubuntu-latest
+    outputs:
+      os: ${{ env.OS }}
+      container: ${{ env.CONTAINER }}
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          repository: teaxyz/pantry.core
+      - uses: teaxyz/setup@v0
+      - run: scripts/get-platform.ts ${{ inputs.projects }}
+        env:
+          PLATFORM: ${{ inputs.platform }}
+
   bottle:
-    runs-on: ${{ matrix.platform.os }}
+    needs: [get-platform]
+    runs-on: ${{ needs.get-platform.outputs.os }}
     defaults:
       run:
         working-directory: tea.xyz/var/pantry
-    strategy:
-      matrix:
-        platform:
-          - os: macos-11
-            name: darwin+x86-64
-          - os: ubuntu-latest
-            name: linux+x86-64
-          - os: [self-hosted, macOS, ARM64]
-            name: darwin+aarch64
-          - os: [self-hosted, linux, ARM64]
-            name: linux+aarch64
     outputs:
       srcs: ${{ env.srcs }}
       built: ${{ env.built }}
@@ -55,10 +62,10 @@ jobs:
       - uses: actions/download-artifact@v3
         if: ${{ inputs.new-version }}
         with:
-          name: ${{ matrix.platform.name }}
+          name: ${{ inputs.platform }}
           path: tea.xyz/var/pantry
 
-      - run: scripts/fetch-pr-artifacts.ts ${{ github.repository }} ${{ github.sha }} ${{ matrix.platform.name }} >>$GITHUB_ENV
+      - run: scripts/fetch-pr-artifacts.ts ${{ github.repository }} ${{ github.sha }} ${{ inputs.platform }} >>$GITHUB_ENV
         if: ${{ !inputs.new-version }}
         env:
           GITHUB_TOKEN: ${{github.token}}
@@ -81,14 +88,14 @@ jobs:
       # FIXME: replace this with a tea script based on https://localazy.com/blog/how-to-automatically-sign-macos-apps-using-github-actions
       # github has a doc with similar content, but it's not returning to me atm.
       - uses: apple-actions/import-codesign-certs@d54750db52a4d3eaed0fc107a8bab3958f3f7494
-        if: matrix.platform.name == 'darwin+aarch64' || matrix.platform.name == 'darwin+x86-64'
+        if: inputs.platform == 'darwin+aarch64' || inputs.platform == 'darwin+x86-64'
         with:
           p12-file-base64: ${{ secrets.APPLE_CERTIFICATE_P12 }}
           p12-password: ${{ secrets.APPLE_CERTIFICATE_P12_PASSWORD }}
 
       # Codesign libs and bins
       - name: Codesign package
-        if: matrix.platform.name == 'darwin+aarch64' || matrix.platform.name == 'darwin+x86-64'
+        if: inputs.platform == 'darwin+aarch64' || inputs.platform == 'darwin+x86-64'
         run: |
           for PKG in ${{ env.relative-paths }}; do
             find /opt/$PKG -name '*.so' -or -name '*.dylib' -print0 | \
@@ -99,7 +106,7 @@ jobs:
 
       # This isn't very informative, but even a no-op is safer than none
       - name: Check codesigning
-        if: matrix.platform.name == 'darwin+aarch64' || matrix.platform.name == 'darwin+x86-64'
+        if: inputs.platform == 'darwin+aarch64' || inputs.platform == 'darwin+x86-64'
         run: |
           for PKG in ${{ env.relative-paths }}; do
             for SIG in `find /opt/$PKG -name '*.so' -or -name '*.dylib'` `find /opt/$PKG/bin -type f`; do
@@ -110,7 +117,7 @@ jobs:
 
       # Needed for self-hosted runner, since it doesn't destroy itself automatically.
       - name: Delete keychain
-        if: always() && matrix.platform.name == 'darwin+aarch64'
+        if: always() && inputs.platform == 'darwin+aarch64'
         run: security delete-keychain signing_temp.keychain
 
       - run: |
@@ -153,7 +160,7 @@ jobs:
       - name: upload artifacts
         uses: actions/upload-artifact@v3
         with:
-          name: ${{ matrix.platform.name }}-bottles
+          name: ${{ inputs.platform }}-bottles
           path: artifacts.tar
           if-no-files-found: error
 
@@ -163,17 +170,6 @@ jobs:
     defaults:
       run:
         working-directory: tea.xyz/var/pantry
-    strategy:
-      matrix:
-        platform:
-          - os: macos-11
-            name: darwin+x86-64
-          - os: ubuntu-latest
-            name: linux+x86-64
-          - os: [self-hosted, macOS, ARM64]
-            name: darwin+aarch64
-          - os: [self-hosted, linux, ARM64]
-            name: linux+aarch64
     steps:
       - uses: actions/checkout@v3
         with:
@@ -192,7 +188,7 @@ jobs:
 
       - uses: actions/download-artifact@v3
         with:
-          name: ${{ matrix.platform.name }}-bottles
+          name: ${{ inputs.platform }}-bottles
 
       - run: |
           tar xvf $GITHUB_WORKSPACE/artifacts.tar
@@ -222,8 +218,6 @@ jobs:
           AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
           AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
 
-      #NOTE ideally we’d invalidate all at once so this is atomic
-      # however GHA can’t consolidate outputs from a matrix :/
       - uses: chetan/invalidate-cloudfront-action@v2
         env:
           PATHS: ${{ steps.upload.outputs.cf-invalidation-paths }}
