+++ .github/workflows/build.yml
@@ -6,38 +6,30 @@ on:
       projects:
         required: true
         type: string
+      platform:
+        required: true
+        type: string
 
 jobs:
-  check-core-sizes:
+  get-platform:
     runs-on: ubuntu-latest
     outputs:
-      gha-linux-build-size: ${{ env.GHA_LINUX_BUILD_SIZE }}
+      os: ${{ env.OS }}
+      container: ${{ env.CONTAINER }}
     steps:
       - uses: actions/checkout@v3
         with:
           repository: teaxyz/pantry.core
       - uses: teaxyz/setup@v0
-      - run: scripts/check-core-sizes.ts ${{ inputs.projects }}
+      - run: scripts/get-platform.ts ${{ inputs.projects }}
         env:
+          PLATFORM: ${{ inputs.platform }}
           TEA_PANTRY_PATH: ${{ github.workspace }}
 
   build:
-    runs-on: ${{ matrix.platform.os }}
-    needs: [check-core-sizes]
-    strategy:
-      matrix:
-        platform:
-          - os: macos-11
-            name: darwin+x86-64
-          - os: ${{ needs.check-core-sizes.outputs.gha-linux-build-size }}
-            name: linux+x86-64
-            container:
-              image: debian:buster-slim
-          - os: [self-hosted, macOS, ARM64]
-            name: darwin+aarch64
-          - os: [self-hosted, linux, ARM64]
-            name: linux+aarch64
-    container: ${{ matrix.platform.container }}
+    runs-on: ${{ needs.get-platform.outputs.os }}
+    needs: [get-platform]
+    container: ${{ needs.get-platform.outputs.container }}
     steps:
       - uses: actions/checkout@v3
 
@@ -89,33 +81,16 @@ jobs:
       - name: upload artifacts
         uses: actions/upload-artifact@v3
         with:
-          name: ${{ matrix.platform.name }}
+          name: ${{ inputs.platform }}
           path: artifacts.tgz
           if-no-files-found: error
 
   test:
-    needs: [build]
-    runs-on: ${{ matrix.platform.os }}
-    strategy:
-      matrix:
-        platform:
-        - os: macos-11
-          name: darwin+x86-64
-        - os: ubuntu-latest
-          name: linux+x86-64
-        - os: ubuntu-latest
-          name: linux+x86-64
-          container: ghcr.io/teaxyz/infuser:latest
-        - os: ubuntu-latest
-          name: linux+x86-64
-          container: debian:buster-slim
-        - os: [self-hosted, macOS, ARM64]
-          name: darwin+aarch64
-        - os: [self-hosted, linux, ARM64]
-          name: linux+aarch64
+    needs: [get-platform, build]
+    runs-on: ${{ needs.get-platform.outputs.os }}
     outputs:
       HAS_SECRETS: ${{ env.HAS_SECRETS }}
-    container: ${{ matrix.platform.container }}
+    container: ${{ needs.get-platform.outputs.container }}
     steps:
       - uses: actions/checkout@v3
       - uses: teaxyz/setup@v0
@@ -125,7 +100,7 @@ jobs:
 
       - uses: actions/download-artifact@v3
         with:
-          name: ${{ matrix.platform.name }}
+          name: ${{ inputs.platform }}
 
       - name: extract bottles
         run: tar xzf artifacts.tgz -C $TEA_PREFIX
@@ -145,17 +120,6 @@ jobs:
     # this only works for PRs from our team to our repo (security! :( )
     if: startsWith(github.ref, 'refs/pull/') && startsWith(github.repository, 'teaxyz/pantry.') && needs.test.outputs.HAS_SECRETS == 'true'
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        platform:
-          - os: macos-11
-            name: darwin+x86-64
-          - os: ubuntu-latest
-            name: linux+x86-64
-          - os: [self-hosted, macOS, ARM64]
-            name: darwin+aarch64
-          - os: [self-hosted, linux, ARM64]
-            name: linux+aarch64
     steps:
       - uses: actions/checkout@v3
       - uses: teaxyz/setup@v0
@@ -165,12 +129,12 @@ jobs:
 
       - uses: actions/download-artifact@v3
         with:
-          name: ${{ matrix.platform.name }}
+          name: ${{ inputs.platform }}
 
       - run: cache-artifacts.ts
           ${{github.repository}}
           ${{github.ref}}
-          ${{matrix.platform.name}}
+          ${{inputs.platform}}
           artifacts.tgz
         env:
           AWS_S3_CACHE: ${{ secrets.AWS_S3_CACHE }}
